refactor: TODO #29 — single T0 definition; VariancePortfolio = reference only
- Payoffs.Log.logPortfolioQ96 is the one T0 (bare log portfolio, continuous
  lnQ96); LadderPosition imports it.
- VariancePortfolio.fromDef6 = N_id·logPortfolioQ96 + R; fromLegs delegates
  (second code path removed); module header labels it T0 reference/diagnostic,
  not a position — the benchmark is T1 (LadderPosition).
- app: panel-replica-vs-hopB retired (units mismatch; superseded by
  panel-t1-vs-t0 / panel-t2-vs-t1).
- test: fromLegs = N_id·logPortfolioQ96 + R exactly at 6 ticks.
- README Def 7 T0 marked reference-only with the code pointer; TODO #29.
